+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 ### Deprecated
 - Datasource: elasticsearch_{destination} in favor of elasticsearch_opendistro_{destination}
 
+### Fixed
+- Using date math in index names - an index resource is tied to the resolved index it is created with.
+
+
 ## [1.2.0] - 2020-05-31
 ### Added
 - Add aws profile authentication option